Remove <scope>test from mycila
This allows odlparent-based projects to use Guice with Mycila more
easily in non-test related Java environments, such as in a pure JSE
opendaylight.simple (POC).

As per the original commit message when Mycila was first added to 
odlparent, there is still no intention of using it (or Guice) in our
standard deployment model of running ODL in OSGi.  Therefore, this 
contribution does still not include an OSGi MANIFEST.MF for Mycila,
nor a Karaf wrap feature for it, as we still do not need that.

Change-Id: I1f5586bd1dcf8d12756e33f4b74c4326b924058d
Signed-off-by: Michael Vorburger <vorburger@redhat.com>
(cherry picked from commit f48bc51f9dd2ae64b16cd94bbaeb1867eaef2934)