Use SpotBugs 3.1.2, and NP_NONNULL_PARAM_VIOLATION
This patch upgrades the SpotBugs plugin and engine to 3.1.2. The
default version of the engine for plugin 3.1.2 is 3.1.1, so we still
need to override that...

SpotBugs 3.1.2 supports the nullable annotations in Checker Framework
(see
https://github.com/spotbugs/spotbugs/commit/dd436ff56abf0273a080ad228fcf5abd8ee3a097
for the patch), so we can re-enable NP_NONNULL_PARAM_VIOLATION when
using SpotBugs. This involves using a separate configuration, so this
patch splits everything out, including the report directory for
projects which run both FindBugs and SpotBugs.

Change-Id: I8cb110f12d79414d476b7a6c525412f3b7e2c188
Signed-off-by: Stephen Kitt <skitt@redhat.com>