ODLPARENT-137: restore the opendaylight prompt
In Karaf 4.1 and later, it turns out that the prompt defined in the
branding is overridden by the shell initialisation script,
etc/shell.init.script, which defines its own prompt. This patch adds
our own version of that script which disables the override.

Thanks to Guillaume Nodet for figuring out where the override was
coming from, and to Michael Vorburger for the initial investigation!
See https://issues.apache.org/jira/browse/KARAF-5570 for details.

Change-Id: I22410cb0425ab5b8a3f6e7c410fbb76587f9f973
Signed-off-by: Stephen Kitt <skitt@redhat.com>