Update patch set 2
Patch Set 2: Code-Review-1

> What kind of impact will this have on our CI?
 > 
 > We currently allocate 40 GB for our build systems.

yeah, this might not be good for CI.

If this kind of change is agreed to be the default, we can always
tweak our CI deployments to be smaller sizes if needed. Or we can
get bigger disks as well, but I think just like this we'll be at
risk of some trouble.

Patch-set: 2
Reviewer: Gerrit User 2264 <2264@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Code-Review=-1