Update patch set 2
Patch Set 2:

> > > What kind of impact will this have on our CI?
 > > >
 > > > We currently allocate 40 GB for our build systems.
 > >
 > > yeah, this might not be good for CI.
 > >
 > > If this kind of change is agreed to be the default, we can always
 > > tweak our CI deployments to be smaller sizes if needed. Or we can
 > > get bigger disks as well, but I think just like this we'll be at
 > > risk of some trouble.
 > 
 > This change doesn’t increase the amount of logs we produce, it only
 > increases the maximum size of any single log file. If our CSIT jobs
 > produce 2GiB of logs, instead of having (2GiB / 16MiB) log files,
 > we’d have one.

But CSIT chokes on large files. Well, it has in the past. So in the case
that we hit a bug which results in really chatty logs, or even just 
a logging mistake bug (we've had both before) that fills the log then
our jobs choke, abort and we get nothing but a red dot and no info.

Patch-set: 2