Update patch set 2
Patch Set 2:

> But CSIT chokes on large files. Well, it has in the past. So in the
 > 
 > case that we hit a bug which results in really chatty logs, or even
 > just a logging mistake bug (we've had both before) that fills the
 > log then our jobs choke, abort and we get nothing but a red dot
 > and no info.

Jamo, and to underline skitt's point, let us understand this: 

so CSIT chokes on a 10 GB file but is OK with 625x (?) 16 MB karaf.log.1 files? ;)

The 16 MB default are *REALLY* low (with our INFO instead of WARN/ERROR default).

Would there be a middle ground that would let people +1 this?  

Say if it was changed from 16 MB to 1, 3 or 5 (instead of 10) GB?

> We configure this to be 10GB by default in puppet-opendaylight FYI

right, that's why I noted in https://bugzilla.redhat.com/show_bug.cgi?id=1570083#c2: "but it's really an indication that the default upstream could be better."

> We also reduce the number of roll-overs from 10 to 2.

Tx; noted in ODLPARENT-153.

Patch-set: 2