Update patch set 2
Patch Set 2: Code-Review-2

So an out-of-box installation of ODL can rapidly consume tens of gigabytes of diskspace?

If a deployment requires more logs to be kept, it should be tweaked in deployment.

Overwhelming operators with INFO-level messages ... where are those coming from? Are the callers compliant with https://wiki.opendaylight.org/view/BestPractices/Logging_Best_Practices ?

Patch-set: 2