Update patch set 2
Patch Set 2:

> If a deployment requires more logs to be kept, it should be tweaked
 > in deployment.

It already is; in CSIT, and in installers such as Triple'O; see ODLPARENT-154.

 > So an out-of-box installation of ODL can rapidly consume tens of
 > gigabytes of diskspace?
 > Overwhelming operators with INFO-level messages ... where are those
 > coming from? Are the callers compliant with
 > https://wiki.opendaylight.org/view/BestPractices/Logging_Best_Practices

perhaps not, and you are welcome to take this up with projects such as netvirt and genius.  That is a little beyond the scope of this proposed changed though, because as per ODLPARENT-154: We consider this default of a tiny 16 MB log size a regression introduced by a recent odlparent change.  It used to be much bigger.  Do you think 16 MB is a suitable default?  

Would there be a middle ground that would let people +1 this?  

Say if it was changed from 16 MB to 1, 3 or 5 (instead of 10) GB?

Patch-set: 2