Update patch set 2
Patch Set 2:

As far back as I can remember, the default was 10x1MiB -- it definitely is true for both Carbon and Nitrogen. So this is definitely a regression in odlparent.

I do not consider anything in gigabyte range to be a reasonable default -- and please note that I am quite used to analyzing 30+GiB of logs.

It is a downstream (as in ODL downstreams) problem that they do not follow best practices which have been agreed to eons ago, leading to poor end user experience. It definitely should not be addressed here.

Patch-set: 2