Update patch set 2
Patch Set 2:

I am sorry, defining common third-party is completely in scope of ODL Parent project, available at https://wiki.opendaylight.org/view/Project_Proposals:ODL_Root_Parent, if you are not familiar with it.

It does not preclude you from using whatever version you like. It's just more painful -- and rightfully so, as if your work is ready to be widely used, it should be considered for movement into odlparent.

That has its cost in development time, but that is a trade-off of stable features. It means that downstream do not have to deal with you rapidly changing things -- and that is right, ODLs focus are its use cases, so downstreams should be disturbed as little as possible.

Guava is relevant, because it is *way* more important to OpenDaylight as a platform than guice, hence the need to propagate versions would by that logic be more important. As it turns out, it is sufficient to upgrade Guava at most once in a release.

As for the question of infrautils: it is creating a separate pom.xml hierarchy, just because of your steady refusal to pick a non-odlparent dependency. Honestly, wouldn't it be much better if the stuff from infrautils were available with 'binding-parent' configuration?

That separate pom.xml hierarchy is already create PITA for me as yangtools PTL, as suddenly I have to propagate yangtools version to genius, which should be getting yangtools through mdsal...

Patch-set: 2