Update patch set 2
Abandoned

This is a PITA and would take more time than I'm willing to sink into it.

Abandoning this during my big clean up before the lazy european ;) summer, and more importantly to reduce my own “mental load”... given that I've not managed to wrap this one up in so long, clearly it's not really all that critical in the bigger picture of things. It could always be Restored, by me - or anyone else interested in picking this up. Onwards. Focused.

Patch-set: 2
Status: abandoned
Tag: autogenerated:gerrit:abandon