Update patch set 1
Patch Set 1: Code-Review-1

If it's just a warning, this isn't blocking, right?  How about keeping the warning as a reminder that we really ought to fix this?  Is it possible to learn more details about this?  What is the html4 vs html5 javadoc, and how could we get the new one for ODL?  May be worth a thread on odlparent-dev?

Patch-set: 1