Update patch set 6
Patch Set 6:

> do we really want to do this?  Won't this have non negligeable
 > impacts and require adapting a lot of code?  Do we have to do it? 
 > Why?  How would we "justify" why we did this - is Mockito 2 that
 > much better than v1 that it's "justified" for us to switch over
 > now?  I vaguely remember having myself proposed this, long ago, and
 > having being shut down (presumably for good reason).

We need to measure the impact. Mockito 2 brings a number of benefits, one of which being that it greatly reduces the need for PowerMock (it can handle more static mocks itself).

Back when we were snapshot-integrated this would have been impossible; for v4 we can at least consider it...

Patch-set: 6
Label: Verified=0