Route test logging through log4j2
Some tests need to capture logging output, which is impossible
to do with slf4j-simple and bringing in logback (or any other
implementation) results in a nasty warning being printed.

Switch over to log4j-slf4j-impl, which implements slf4j via log4j2,
which searches for configuration when initialized. Normal users
will not provide a configuration, hence output will be redirected
to console, just as with slf4j-simple.

If a test needs to capture logging output, it will provide the
configuration on its classpath, leading to output being routed
as needed.

Change-Id: I66246542dccd60f660f67f16326d91b49dbcbf7a
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>