Update patch set 1
Patch Set 1: Code-Review+1

(1 comment)

This makes sense to me as a basis for https://git.opendaylight.org/gerrit/#/c/75066/ or now rather https://git.opendaylight.org/gerrit/#/c/61499/, which I've now built on top of this.  Giving a +1 instead of a +2 and merge while awaiting the Robert's +1 on c/61499 and Abandon of c/75066.

Patch-set: 1
Label: Code-Review=+1