Remove blueprint plugin generatedDir
This customization was needed for planned extended behavior
in controller's blueprint extension, which never materialized.

Controller can reintroduce this configuration with the new plugin,
if the need arises again -- remove it to reduce our clutter.

We also remove Import-Service and Export-Service headers from
the manifest, as they are deprecated and we have not used them
anywhere.

Change-Id: I61560793c4631f24d9ba353c7d9242ae5837e821
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>