Add odl-apache-sshd feature
This is packaging of Apache SSHD, which needs to be co-managed
with karaf, as it is also using it. At this point karaf and us
are separated by a major version, so this is a simple exercise.

JIRA: ODLPARENT-158
Change-Id: Iceef03542bfba365a968572b58bef2d6c7ed1018
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>
(cherry picked from commit b612b2eb27297d5f32f4c215f250fa6803963b95)