Update patch set 4
Patch Set 4:

I converted the toaster to use OSGI-INF/blueprint. The bundle fails to resolve b/c the maven-bundle-plugin adds an Import-Service header to the manifest for all service interfaces reference in <reference> elements but there's no bundle contain ing an Export-Service header. Presumably every ODL project that advertises services, specifically the kernel, would have to first be converted to use OSGI-INF/blueprint. 

From what I've read, Import-Service/ Export-Service is actually deprecated and has been for a while - not sure why maven-bundle-plugin still generates it. Another option is outlined in https://stackoverflow.com/questions/18504379/how-to-configure-maven-bundle-plugin-to-exclude-import-service-and-export-servic - I tried _removeheaders and it worked. We could add this to odlparent.

Patch-set: 4