Update patch set 5
Patch Set 5:

> Ah, neat point. The thread at https://dev.eclipse.org/mhonarc/lists/equinox-dev/msg01088.html
 > provides some more insight.
 > 
 > For services we already have features to express their dependency
 > (once they get documented!). Hence I would propose adding:
 > 
 > <_removeheaders>Import-Service,Export-Service</_removeheaders>
 > 
 > to our default bundle-plugin configuration.

We need to do this. Export-Service is only added by the maven-bundle-plugin for BP <service> elements  - it isn't added for services advertised in code like mdsal-binding-dom-adapter does, in that case we would have to manually add the Export-Service headers. I don't think that's desirable. 

So can we add the <_removeheaders> with this patch?

Patch-set: 5