Decouple sonar and jacoco
Our current configuration forces Sonar to consider coverage only
if it occurs within the same module, with downstreams having no easy
way to provide aggregated execution view for sonar.

This patch decouples jacoco and sonar configuration, so that UTs report
their execution profile as usual, but the individual profiles are merged
by the aggregator pom -- and when sonar runs we point it to the
aggregated report.

It also moves the pluginManagement for jacoco-maven-plugin into
odlparent-lite, as aggregators need to invoke it.

Change-Id: Ie590dd98da003c1070c9db06fff90ffed8c065a5
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>