Update patch set 2
Patch Set 2: Code-Review-1

Is this a SpotBugs problem, or a FindBugs problem? I suspect the former, since we don’t run into the issue on FindBugs. If it concerns SpotBugs, this should go into spotbugs-exclude.xml instead.

(I don’t agree with all the points raised in the upstream discussion, but this is too complex an issue to get right with SpotBugs as it currently stands anyway — and as your comment says, good API design would typically involve value objects which aren’t extended in such a way that this becomes an issue. Then there’s the whole topic of proxies and mocks...)

Patch-set: 2
Label: Code-Review=-1