Support building without zip archives
The build can be configured to skip zip or tar.gz archives (or even
both). Currently if zip archives are disabled, the build fails since
features-test expects a zip archive; this patch allows either a zip or
a tar.gz to be used, depending on what's built.

If both are disabled the build will fail unless it can retrieve a
matching archive from a repository. The nominal configuration is still
to build both.

JIRA: ODLPARENT-174
Change-Id: I858810767bcca452eebf52107ee76be129f85192
Signed-off-by: Stephen Kitt <skitt@redhat.com>