Avoid pulling old versions
This uses karaf-plugin’s ability to resolve patched features to
upgrade the versions of javax.mail, commons-beanutils, commons-codec
and aries.util pulled in by Karaf and Pax.

We need to change the order of steps in karaf4-parent so that the
Karaf assembly is processed first, then the resources overwritten with
our versions, before karaf-plugin resolves the dependencies.

JIRA: ODLPARENT-189
Change-Id: Id78d7301bd402bc6872fe8a696af9ddc734cbe8d
Signed-off-by: Stephen Kitt <skitt@redhat.com>
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>