Avoid pulling in old javax.mail and commons-codec
This uses karaf-plugin’s ability to resolve patched features to
upgrade the versions of javax.mail and commons-codec pulled in by
Pax. We need to change the order of steps in karaf4-parent so that the
Karaf assembly is processed first, then the resources overwritten with
our versions, before karaf-plugin resolves the dependencies.

Change-Id: Id78d7301bd402bc6872fe8a696af9ddc734cbe8d
JIRA: ODLPARENT-189
Signed-off-by: Stephen Kitt <skitt@redhat.com>