Allow multi-version builds
Since we are defining maven.compiler.* properties, there is no
point in hard-coding that in plugin configuration, as it will be
picked up automatically.

When we are running with JDK9+, also define maven.compiler.release
to pass --release flag to javac, which fixes cross-compilation
warnings about boot class path -- see JEP-247 for details

Since the version configuration is driven by properties, downstreams
now can easily override the defaults and use that to create JEP-238
jars.

Change-Id: I36e7aaa857b91f1d68eac800c275306325a96e0f
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>