Make javadoc HTML5 support an opt-out feature
Our long-term goal is to support HTML5 javadoc, which is not supported
by our current JDK8 tool chain, but is available with JDK9+.

JDK9+ allow fallowing back to html4 with the appropriate option, but
this option causes JDK11+ to emit a warning.

The transition to HTML5 should not be too hard and we want migrate
seamlessly while still allowing downstreams to defer the migration.

This patch makes the -html4 profile conditional on the existence of
odl.javadoc.html5.optout file in maven project root directory, thus
achieving the goal of not emitting warnings on compliant projects
with newer toolchain.

Change-Id: Ida91e9ffbe74afaac2a2edd43e066e57f73991a4
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>