Remove features-test's exclusions
opendaylight-karaf-empty has three additional dependencies, which
need to be present in local repository when we run SFT. This would
normally be achieved through transitive dependencies, but it does
not work due to features-test exclusion of these.

This was due to infrautils using features-test and also requiring
no classpath duplicates, hence single-feature-parent needed to
explicitly name them -- which broke when we added BouncyCastle
framework extensions and did not add them to single-feature-parent.

This has changed since then, though, and infrautils is using
bundles-test-lib, which does not bring the duplicates, and thus
features-test is used exclusively by SFT.

Therefore remove the exclusion and duplication single-feature-parent,
restoring sanity.

Change-Id: I167bae2965706406b5040a699929963c66e59e31
JIRA: ODLPARENT-130
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>