Drop modernizer back to Java 8
Modernizer produces a ton of recommendations to use List.of()
and similar -- which requires careful consideration, as we are
using ImmutableList.copyOf() all over the place and Java's version
does not offer a check for immutability.

This is a stop-gap until we figure out a way forward.

JIRA: ODLPARENT-201
JIRA: ODLPARENT-203
Change-Id: I7037e6ef09f132129ac3b63145235eb6903961bc
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>