Add leveldbjni-all repackaging
We are publishing all our artifacts to Maven Central, but this jar
has been seeded into our Nexus manually after upstream failed to
make a public release of the contribution for years.

This jar is not present in Central and hence most our features do
not work out of the box, as you still need to point to ODL Nexus
to get this one jar.

This patch rectifies that situation by re-packaging this jar in
org.opendaylight.odlparent namespace and referencing it in our
features.

Change-Id: I16895d6e9e2572ab85db9c509e06306c7acc441d
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>