Make sure CLI components are scope=provided
We are filtering dependencies, but if downstreams fail to set these
dependencies to provided we'll end up pull in transitive dependencies
are required -- and those are provided by the platform. Mark CLI
components as provided, as that is what they are.

Change-Id: I32c6f4d3b24591d828d808c13849907176fd0f99
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>
(cherry picked from commit 04d822f807a47120e3c7267a55a1ada65efe9d7e)