Add default repository to defaultRepositories
SFT reconfigures the repository layout such that Karaf ends up
ignoring its default system repository, which contains various
edits we are doing for the baseline distribution.

This leads to SFT running actually against vanilla Karaf
definitions, which is not what we want.

We need to override the repository config to ensure we pick up
the maven repository, so we can find things like pax-exam, etc.

Fix this by setting org.ops4j.pax.url.mvn.defaultRepositories
to point to karaf.default.repository, so that it is consulted
*before* all other repositories.

JIRA: ODLPARENT-209
Change-Id: I6c0fe2716ca1f4aaf695f20dc142838e13cc9e81
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>