Remove guice dependencies
Guice is a dead upstream and it is being declared here with even
deader mycilla (jsr250).

In ODL world, this integration lives in infrautils and all users
are depending on infrautils anyway. Remove the declaration in
odlparent and let infrautils manage it -- making it possible to
peel it off into a separate project with its own rules.

JIRA: ODLPARENT-212
Change-Id: I8bab3689517b74e6221323344efbb0ea0057bdb5
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>