WIP: Teach SFT about loading test dependencies
Our overall project feature structure has a number of features
which cannot be tested in their home project because the default
implementation of a service they require comes from a downstream
project.

Teach SingleFeatureTest to perform a bulk installation of not
only the feature being tested, but also any features marked with
scope=test.

// FIXME: add two features to actually test the capability

Change-Id: Id7568e1e3840b925a84f91d223dbd6c76146771f
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>