Add org.osgi.service.component.annotations declaration
When trying to have combined JPMS + OSGi artifact, where both are
used to the full extent we fail to compile.

In order to side-step that, such artifacts need to pull in
component annotations and depend on its derived name.

Add a declarations so downstreams do not need to maintain versions
themselves.

Change-Id: If9a4e93a2292578837011132e0f1b2a68d9d347b
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>