WIP: Bump karaf to 4.3.0
Release notes:
https://issues.apache.org/jira/secure/ReleaseNote.jspa?projectId=12311140&version=12343304

This forces us to deal with jakarta.activation-api, as that is no
longer provided out of the box. We deal with that by providing our
own feature and ditching javax.activation declaration, hence we now
have odl-jakarta-activation-api.

We also upgrade to jackson-2.11.3, which finally puts an end to
all those serialization widgets vulns.

Change-Id: Ib34abcbf6948664e5c1f9381fdc8577dc4f62108
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>