Move maven-compiler-plugin declaration
While we would usually want to define plugins at the place of their
first use, maven-compiler-plugin is special do to its
maven.compiler.release directive, which has a direct tie-in with
maven-enforcer-plugin.

This also allows us to compile plugins which we want to use in
odlparent. Yeah, I know, but fixing this odlparent-lite/odlparent split
is a large-scale effort, which really is on back burner.

So let's do the minimal thing of relocating basic compiler plugin
declaration into odlparent-lite.

Change-Id: I39943bb3308ddf1b1f8447ce869a52c22b8283cf
JIRA: ODLPARENT-220
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>