Propagate JaCoCo argLine to Karaf as needed
We have disabled JaCoCo execution reports because they do not work
with SFT correctly. This problem boils down to running the right JVM
with JaCoCo options -- it should not be the surefire JVM, but rather
Karaf we start from it.

Use @{argLine} special handling in surefire to pass the argLine (set by
JaCoCo) as a system property rather than it being a JVM argument.

This does not solve integration with Sonar, as we still need to deal
with the XML report being empty, but it is a start and an example for
others using pax-exam.

JIRA: ODLPARENT-274
Change-Id: I9b33e5b5c9b3b3b2b347f4f4ef034acb677d501a
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>
(cherry picked from commit e474188bbc706917d2966928cba0283d60f989e6)