Propagate maven.compiler.release to xtend-maven-plugin
xtend picks up the target version from maven.compiler.source, which we
are not setting since our migration to Java 11. This results in a
slightly sub-optimal code, as we end up generating Java snippets for
Java 6 -- and not using lambdas, switch on strings and similar.

Fix this by explicitly configuring for maven.compiler.release, so we
pick up our actual target.

Change-Id: I99fb0748967f42bbbc4e01ec763faa081127471a
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>