Delay cyclonedx-maven-plugin
We really should execute as the last plugin in package phase, but that
may not work due to configuration inheritence and other plugins butting
in.

We really would like to execute before install or deploy runs, which
would seem to indicate 'verify', but that's not exactly true, as
'verify' step should be looking at the BOMs we produced and make sure
we capture everything that ends up installed/deployed.

The best option for developer convenience is post-integration-test, we
will not be executed if integration tests fail.

Change-Id: I181e564b7ecf287140aea731f74a16eba8b5bf35
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>