Disable CT_CONTRUCTOR_THROW
This warning comes up in surprising contexts and sometimes it can be
hard to diagnose what exactly is triggering it.

Polluting classes with finalize() is a non-option. Being overly-clever
with constructors is a non-starter as well. Sprinkling
@SuppressFBWarnings on code which has been stable for years also does
not make sense.

Disable this warning, as its relevance is rapidly diminishing as we
require Java versions beyond 18.

Change-Id: Ic1532d4051fb76e1b2eeae66d49342b0e9c3f176
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>