Remove unnecessery osgi version processing
The osgiVersion method is unnecessary because the result is then
passed to the Feature#setVersion() method, which also processes the version
cleanup using the VersionCleaner.clean() method.

This method also replaces the dash used to divide qualifier and micro
version in ODL SNAPSHOT version with a dot.

Replace our cleanup with the VersionCleaner.clean() method
in the feature version taken from the dependency.

JIRA: ODLPARENT-315
Change-Id: I3979f001f7ad57557a4208f9a9cbdc2c34f29d8b
Signed-off-by: Peter Suna <peter.suna@pantheon.tech>